Admin module (WIP)

+++ routes/web.php
@@ -1,5 +1,7 @@
 <?php
 
+use LaraToDo\User;
+use Illuminate\Support\Facades\Auth;
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -12,47 +14,66 @@
 */
 
 Route::group(['middleware' => 'auth'], function(){
-	Route::get('/index', 'Auth\LoginController@index');
-	Route::get('login', function(){
-		return view('auth/login');
-	});
+	if(Auth::user() == User::where('roles', 'admin')){
+		Route::get('/admin/', 'AdminController@show');
+		Route::post('/admin/edit', 'EditProfileController@edituser');
+		Route::get('/admin', function(){
+			return view('admin');
+		});
+		Route::post('/checklogin', 'Auth\LoginController@checklogin')->name('checklogin');
+		Route::get('/logout', 'Auth\LoginController@logout')->name('auth.logout');
 
-	Route::post('/tasks/{task}/done', 'TaskController@done');
-	Route::post('/tasks/{task}/undone', 'TaskController@undone');
-	Route::delete('/tasks/{task}', 'TaskController@destroy');
+		Route::get('login', function(){
+			return view('auth/login');
+		});
 
-	Route::get('/tasks/', 'TaskController@show');
-	Route::post('/tasks', 'TaskController@store');
+		Route::get('/', function(){
+			Auth::logout();
+			return view('welcome');
+		});
+	}
+	elseif(Auth::user() == User::where('roles', 'user')){
+		Route::post('/checklogin', 'Auth\LoginController@checklogin')->name('checklogin');
+		Route::get('/logout', 'Auth\LoginController@logout')->name('auth.logout');
 
-	Route::put('/tasks/', 'TaskController@update');
+		Route::get('login', function(){
+			return view('auth/login');
+		});
 
-	Route::post('/user/edit', 'EditProfileController@edituser');
+		Route::get('/', function(){
+			Auth::logout();
+			return view('welcome');
+		});
 
-	Route::get('/edit', function(){
-		return view('edituser');
-	});
+		Route::get('/index', 'Auth\LoginController@index');
+		Route::post('/tasks/{task}/done', 'TaskController@done');
+		Route::post('/tasks/{task}/undone', 'TaskController@undone');
+		Route::delete('/tasks/{task}', 'TaskController@destroy');
 
-	Route::get('resetpassword', function(){
-		return view('auth/passwords/reset');
-	});
+		Route::get('/tasks/', 'TaskController@show');
+		Route::post('/tasks', 'TaskController@store');
 
-	Route::post('/resetpass', 'Auth\ResetPasswordController@reset');
+		Route::put('/tasks/', 'TaskController@update');
+
+		Route::post('/user/edit', 'EditProfileController@edituser');
+
+		Route::get('/edit', function(){
+			return view('edituser');
+		});
+
+		Route::get('resetpassword', function(){
+			return view('auth/passwords/reset');
+		});
+
+		Route::post('/resetpass', 'Auth\ResetPasswordController@reset');
+	}
+
 });
-
+	
+	Route::post('/checklogin', 'Auth\LoginController@checklogin')->name('checklogin');
+	
 	Route::get('/about', function () {
 	    return view('about');
-	});
-
-	Route::post('/checklogin', 'Auth\LoginController@checklogin')->name('checklogin');
-	Route::get('/logout', 'Auth\LoginController@logout')->name('auth.logout');
-
-	Route::get('login', function(){
-		return view('auth/login');
-	});
-
-	Route::get('/', function(){
-		Auth::logout();
-		return view('welcome');
 	});
 
 	Route::get('register', function(){
